requirements.txt
torch
torchvision
pandas
matplotlib
numpy<2
scikit-learn
openai
tiktoken
gpt-cost-estimator
loguru
ipykernel
tqdm
pytesseract
pillow
evaluate
jiwer
requests
tabulate
python-dotenv
boto3
transformers
diskcache
google-cloud-vision
omegaconf